These are some changes to make tastudio work for me the way I use it. Some of them, for example the way to decide which states to save and which to throw away are hacks that need to be done better.
Change MovieSession to make it so that the toggling of read only mode controls whether input comes from the user or from the movie.

Hack TasStateManager to for sure keep the last 100 states prior to the current frame. This could be done better, but effectively we want every state in the area where the user is working so that going forward and back a few frames at a time is smooth.

change MainForm to tell the movie to go to play mode when the user switches to read only.

Tastudio add a recursion check to the rewind handler

Tastudio navigation change so that the current frame is visible